rag-app-resumes/requirements.txt
#aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
attrs==23.2.0
azure-common==1.1.28 
azure-core # ==1.29.7 # Azure Core library for shared functionality
azure-identity #==1.15.0 # Azure Identity library for authentication
azure-search-documents #==11.4.0b11 # Client library for Azure Cognitive Search
azure-storage-blob
blis==0.7.11
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.2.1
confection==0.1.4
# contourpy==1.2.0
cryptography==42.0.1
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.3
debugpy==1.8.0
decorator==5.1.1
distro==1.9.0
executing==2.0.1
fonttools==4.47.2
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
ipykernel==6.29.0
ipython==8.20.0
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.1.17
langcodes==3.3.0
langsmith==0.0.84
MarkupSafe==2.1.4
marshmallow==3.20.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
msal==1.26.0
msal-extensions==1.1.0
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.3
openai==1.10.0
packaging==23.2
parso==0.8.3
pillow==10.2.0
platformdirs==4.1.0
portalocker==2.8.2
preshed==3.0.9
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
pypdf==4.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
#pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
SQLAlchemy==2.0.25
srsly==2.4.8
stack-data==0.6.3
tenacity==8.2.3
thinc==8.2.2
tiktoken==0.5.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.2.0
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
yarl==1.9.4
opentelemetry-api # for tracing
azure-ai-projects # client library for using the Azure AI service
azure-ai-inference[prompts] # prompting tools for the Azure AI service
opentelemetry-sdk # for tracing
opentelemetry-instrumentation # for tracing
opentelemetry-instrumentation-logging # for logging
azure-monitor-opentelemetry # for exporting telemetry to Azure Monitor
azure-monitor-opentelemetry-exporter
flask